+++ src/client/components/Offline/index.jsx
@@ -1,0 +1,3 @@
+import Offline from './Offline'
+
+export default Offline
