+++ src/client/components/Offline/offline.css
@@ -1,0 +1,3 @@
+.offlineWrapper {
+    text-align: center;
+}
